+++ src/controllers/role.js
@@ -1,4 +1,4 @@
-const User = require('../models/User');
+const User = require('../models/GoogleUser');
 
 const decideRole = async (req, res) => {
   try {
