+++ src/passport-configuration/passport-google-setup.js
@@ -8,7 +8,7 @@ passport.use(
     {
       clientID: process.env.GOOGLE_CLIENT_ID,
       clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-      callbackURL: 'http://localhost:3000/api/auth/google/callback',
+      callbackURL: 'http://localhost:3000/api/googleAuth/google/callback',
     },
     async (accessToken, refreshToken, profile, done) => {
       try {
